refactor(MainPage.vue): 자동로그인 모든 페이지에서 동작하도록 코드 main.js로 이동
