+++ frontend/src/main.js
@@ -3,6 +3,8 @@ import App from "./App.vue";
 import vuetify from "./plugins/vuetify";
 import router from "./router";
 import store from './store'
+import {mapMutations} from "vuex";
+import {SET_MEMBER} from "@/store/shared/mutationTypes";
 
 Vue.config.productionTip = false;
 
@@ -10,5 +12,24 @@ new Vue({
   vuetify,
   router,
   store,
+  async created() {
+    const accessToken = localStorage.getItem("token");
+
+    if (accessToken !== null) {
+      const member = await fetch("http://localhost:8080/members/me", {
+        method: 'GET',
+        headers: {
+          'Authorization': 'Bearer ' + accessToken
+        }
+      });
+      if (member.ok) {
+        const memberInfo = await member.json();
+        this.setMember(memberInfo);
+      }
+    }
+  },
+  methods: {
+    ...mapMutations([SET_MEMBER]),
+  },
   render: h => h(App)
 }).$mount("#app");
